move `ImageType` to a more suitable location
