+++ supervision/key_points/annotators.py
@@ -6,7 +6,7 @@ from logging import warn
 import cv2
 import numpy as np
 
-from supervision.annotators.base import ImageType
+from supervision.draw.base import ImageType
 from supervision.detection.utils.boxes import pad_boxes, spread_out_boxes
 from supervision.draw.color import Color
 from supervision.draw.utils import draw_rounded_rectangle
